requirements.txt
Django==3.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0